requirements.txt
ultralytics>=8.0.0
avalanche-lib>=0.3.1
torch>=1.10.0
torchvision>=0.11.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
numpy>=1.20.0
Pillow>=8.0.0
kaggle>=1.5.12

